Added in new CVAR for map swap protection
Since match swap is controlled by admin status, in situations where we want to be able to allow anyone to swap maps (practice/scrims) it should have an override to let it be unprotected specifically if needed.

As such added new CVAR
matchzy_protect_map_swap

When set to true will allow only admins to swap maps, while if set to false will allow anyone to swap maps as long as the match has not started yet.